# openapi.yaml
openapi: 3.0.1
info:
  title: My API
  description: Документация к API проекта CompClub
  version: v1
paths:
...
  /api/Booking/create_advanced_booking:
    post:
      tags:
        - Booking
      summary: "Создание предварительного бронирования (одного активного на аккаунт).\r\nДоступно только для клиентов."
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdvancedBookingModel'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateAdvancedBookingModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateAdvancedBookingModel'
      responses:
        '200':
          description: OK
...
  /api/Booking/get_booking/{id}:
    get:
      tags:
        - Booking
      summary: "Получение информации о конкретном бронировании по ID.\r\nДоступно для владельцев, администраторов и маркетологов."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
...
  /api/Booking/delete/{id}:
    delete:
      tags:
        - Booking
      summary: Удаление конкретного бронирования клиента по ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
  /api/Booking/get_info/{id}:
    get:
      tags:
        - Booking
      summary: Получение информации о конкретном бронировании текущего клиента по ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
...
  /api/Equipment/create_equipment:
    post:
      tags:
        - Equipment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEquipmentModel'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateEquipmentModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateEquipmentModel'
      responses:
        '200':
          description: OK
  /api/Equipment/info/{id}:
    get:
      tags:
        - Equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
  /api/Equipment/update/{id}:
    put:
      tags:
        - Equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipment'
          text/json:
            schema:
              $ref: '#/components/schemas/Equipment'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Equipment'
      responses:
        '200':
          description: OK
  /api/Equipment/delete/{id}:
    delete:
      tags:
        - Equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
...
  /api/Equipment/create_equipment_maintenance:
    post:
      tags:
        - Equipment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentMaintenance'
          text/json:
            schema:
              $ref: '#/components/schemas/EquipmentMaintenance'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EquipmentMaintenance'
      responses:
        '200':
          description: OK
  /api/Equipment/update_equipment_maintenance/{id}:
    put:
      tags:
        - Equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentMaintenance'
          text/json:
            schema:
              $ref: '#/components/schemas/EquipmentMaintenance'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EquipmentMaintenance'
      responses:
        '200':
          description: OK
...
  /api/Tariff/create_tariff:
    post:
      tags:
        - Tariff
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tariff'
          text/json:
            schema:
              $ref: '#/components/schemas/Tariff'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Tariff'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Tariff'
            application/json:
              schema:
                $ref: '#/components/schemas/Tariff'
            text/json:
              schema:
                $ref: '#/components/schemas/Tariff'
  /api/Tariff/update_tariff/{id}:
    post:
      tags:
        - Tariff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tariff'
          text/json:
            schema:
              $ref: '#/components/schemas/Tariff'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Tariff'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Tariff'
            application/json:
              schema:
                $ref: '#/components/schemas/Tariff'
            text/json:
              schema:
                $ref: '#/components/schemas/Tariff'
  /api/Tariff/delete_tariff/{id}:
    delete:
      tags:
        - Tariff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Tariff'
            application/json:
              schema:
                $ref: '#/components/schemas/Tariff'
            text/json:
              schema:
                $ref: '#/components/schemas/Tariff'
...
components:
  schemas:
...
    CreateAdvancedBookingModel:
      type: object
      properties:
        idWorkingSpace:
          type: integer
          format: int32
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
      additionalProperties: false
...
    CreateEquipmentModel:
      type: object
      properties:
        type:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        specification:
          type: string
          nullable: true
        purchasePrice:
          type: number
          format: double
        idClub:
          type: integer
          format: int32
      additionalProperties: false
...
    Equipment:
      type: object
      properties:
        id:
          type: integer
          format: int32
        type:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        specification:
          type: string
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        purchasePrice:
          type: number
          format: double
          nullable: true
        idClub:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    EquipmentMaintenance:
      type: object
      properties:
        id:
          type: integer
          format: int32
        equipmentId:
          type: integer
          format: int32
        maintenanceDate:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
          nullable: true
        cost:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
...
    Tariff:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        pricePerMinute:
          type: number
          format: double
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: Введите 'Bearer' [пробел] и ваш токен в поле ниже для доступа к защищенным ресурсам
      name: Authorization
      in: header
security:
  - Bearer: []